[Feat] skeleton with lazy, suspense
