+++ src/pages/RollingPage/Messages.jsx
@@ -1,37 +1,17 @@
-//messages 를 복사한 페이지 입니다.
+import { Suspense, lazy } from "react";
 import * as S from "./RollingPage.style";
-import PostCardUI from "../../components/RollingCard/RollingCard";
 import PlusIcon from "../../assets/icons/PlusIcon.svg";
 import { Link } from "react-router-dom";
-import { useEffect, useState } from "react";
-import Skeleton from "./Skeleton/Skeleton";
+import Skeleton from "./Skeleton/Skeleton.jsx";
 //
+const DynamicContent = lazy(() => import("./Messages.data.jsx")); // ✨ 데이터 의존 부분만 분리
 
 export default function Message({
   isEdit,
   messages,
   deletedIds,
   setDeletedIds,
-  isLoading,
 }) {
-  const [filterMessages, setFilterMessages] = useState(messages);
-
-  const handleClickFilter = (e) => {
-    const value = e.currentTarget.dataset.value;
-    if (!value) return;
-    const disappear = filterMessages.filter((message) => message.id != value);
-    setFilterMessages(disappear);
-
-    setDeletedIds((prev) =>
-      deletedIds.includes(value)
-        ? deletedIds.filter((item) => item !== value)
-        : [...prev, value]
-    );
-  };
-
-  useEffect(() => {
-    setFilterMessages(messages);
-  }, [messages]);
   //
   return (
     <S.GridBoxes>
@@ -42,21 +22,14 @@ export default function Message({
           </Link>
         </S.CreateBox>
       )}
-      {isLoading ? (
-        <Skeleton />
-      ) : (
-        filterMessages.map((message) => {
-          return (
-            <PostCardUI
-              onClick={handleClickFilter}
-              isEdit={isEdit}
-              key={message.id}
-              data={message}
-            />
-          );
-        })
-      )}
+      <Suspense fallback={<Skeleton />}>
+        <DynamicContent
+          deletedIds={deletedIds}
+          setDeletedIds={setDeletedIds}
+          isEdit={isEdit}
+          messages={messages}
+        />
+      </Suspense>
     </S.GridBoxes>
   );
 }
-//
